backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
bcrypt
PyJWT
python-dotenv
python-multipart
google-genai
httpx
